requirements.txt
opensearch-py==2.4.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0